refactor: projects - 아이콘 크기 변경, experience - UX향상을 위해 정적 레이아웃으로 변경
