+++ src/styles/experiencepage.scss
@@ -10,10 +10,10 @@
 }
 
 .experience-page-middle-container {
-  width: 100%;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
+  width: 50%;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 2%;
   box-sizing: border-box;
   position: relative;
 }
@@ -35,15 +35,14 @@
 }
 
 .experience-page-box {
-  width: 100%;
-  // padding: 5%;
   display: flex;
-  flex-shrink: 0;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
   color: $midnightBlack;
-  font-size: 20px;
+  font-size: 16px;
+  padding: 20px;
+  // border: 2px solid $pastelLilac;
+  // border-radius: 10px;
 }
 
 .experience-page-title {
@@ -54,28 +53,33 @@
 }
 
 .experience-page-subtitle {
-  width: 100%;
   color: $charcoalGray;
-  font-size: 24px;
+  font-size: 20px;
   font-weight: 600;
   margin-bottom: 20px;
-  // margin-bottom: 30px;
 }
 
 .experience-page-contents {
-  width: 50%;
   display: flex;
-  align-items: center;
+  gap: 20px;
+  flex-direction: column;
   justify-content: center;
-  // justify-content: space-between;
 }
 
 .experience-page-image-box {
   display: flex;
   flex-direction: column;
-  img {
-    width: 90%;
-    height: 90%;
+}
+
+.experience-page-link-container {
+  display: flex;
+  gap: 20px;
+  color: $midnightIndigo;
+  a {
+    background-color: $lightpastelLilac;
+    border: none;
+    border-radius: 10px;
+    padding: 0 10px;
   }
 }
 
@@ -83,6 +87,7 @@
   display: flex;
   flex-direction: column;
   line-height: 3;
+  gap: 20px;
   a {
     font-weight: 600;
     &:hover {
@@ -90,7 +95,7 @@
     }
   }
   p {
-    font-size: 16px;
+    font-size: 14px;
     white-space: pre-line;
   }
 }
