+++ src/styles/projectspage.scss
@@ -29,8 +29,8 @@
     }
   }
   svg {
-    width: 40px;
-    height: 40px;
+    width: 30px;
+    height: 30px;
   }
   margin-bottom: 40px;
 }
